+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.2.0 - Unreleased
+
+* Renamed `transducers\vec()` to `transducers\to_traversable()`.
+* Renamed `transducers\is_iterable()` to `transducers\is_traversable()`.
+* Added `transducers\to_fn()` so that transducers can be used with existing
+  reduce functions like `array_reduce`.
+
 ## 0.1.0 - 2014-12-03
 
 Initial release.